Bugs/remaned components (#1584)
* -- Removed the Subscription Error

* renamed NestedAlertContent files

* --Fixed Sample code download issues

* --fixed the project link to the card after create the credential

* --Fixed the checkmark issues on the picker in the return flow project details

* Fixed scopes details as dynamic

* --Separated the Service and formfields page

---------
