--- packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/NestedAlertContentNotSignUp.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { NestedAlertContentEdgeCase } from './NestedAlertContentEdgeCase';
-
-const NestedAlertContentNotSignUp = ({ content }) => {
-  return (
-    <>
-      <NestedAlertContentEdgeCase content={content} isNotSignUp={true}/>
-    </>
-  );
-};
-
-export { NestedAlertContentNotSignUp };
