--- packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/NestedAlertContentType1User.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { NestedAlertContentEdgeCase } from './NestedAlertContentEdgeCase';
-
-const NestedAlertContentType1User = ({content}) => {
-  return (
-    <>
-      <NestedAlertContentEdgeCase content={content} />
-    </>
-  );
-};
-
-export { NestedAlertContentType1User };
