+++ packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/OrganizationAccessDetailsNoProduct.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { OrganizationAccessDetailsEdgeCase } from "./OrganizationAccessDetailsEdgeCase";
+
+const OrganizationAccessDetailsNoProduct = ({content}) => {
+
+  return (
+    <>
+      <OrganizationAccessDetailsEdgeCase content={content} isNoProduct={true} />
+    </>
+  );
+};
+
+export { OrganizationAccessDetailsNoProduct };
