+++ packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/OrganizationAccessDetailsNotMember.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { OrganizationAccessDetailsEdgeCase } from './OrganizationAccessDetailsEdgeCase';
+
+const OrganizationAccessDetailsNotMember = ({content}) => {
+
+  return (
+    <>
+      <OrganizationAccessDetailsEdgeCase content={content} />
+    </>
+  );
+};
+
+export { OrganizationAccessDetailsNotMember };
