+++ packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/OrganizationAccessDetailsNotSignUp.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { OrganizationAccessDetailsEdgeCase } from './OrganizationAccessDetailsEdgeCase';
+
+const OrganizationAccessDetailsNotSignUp = ({ content }) => {
+  return (
+    <>
+      <OrganizationAccessDetailsEdgeCase content={content} isNotSignUp={true}/>
+    </>
+  );
+};
+
+export { OrganizationAccessDetailsNotSignUp };
