+++ packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/OrganizationAccessDetailsType1User.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { OrganizationAccessDetailsEdgeCase } from './OrganizationAccessDetailsEdgeCase';
+
+const OrganizationAccessDetailsType1User = ({content}) => {
+  return (
+    <>
+      <OrganizationAccessDetailsEdgeCase content={content} />
+    </>
+  );
+};
+
+export { OrganizationAccessDetailsType1User };
