+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnAccessToken.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { AccessToken } from '../Card/AccessToken';
 
-const ReturnAccessToken = ({ accessToken,response }) => {
-  return <AccessToken accessToken={accessToken} response={response} />
+const ReturnAccessToken = ({ accessToken,response, scopesDetails }) => {
+  return <AccessToken accessToken={accessToken} response={response} scopesDetails={scopesDetails} />
 }
 
 export { ReturnAccessToken }
